requirements.txt
fastapi==0.82.0
Pillow==9.2.0
requests==2.28.1
uvicorn==0.18.3
